+++ server/src/routes/auth/authRouteHandler.ts
@@ -1,0 +1,10 @@
+import { Router } from "express";
+import authEmailRoute from "./email";
+
+const authRouter = Router();
+
+
+authRouter.use(authEmailRoute);
+
+
+export default authRouter;
